Split pure code from settings
