+++ config/common/add-server.js
