+++ config/common/style-loader-content.js
