+++ config/common/style-loader.js
