+++ config/local.config.js
@@ -2,11 +2,12 @@ import ExtendedDefinePlugin from 'extended-define-webpack-plugin';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import LiveReloadPlugin from 'webpack-livereload-plugin';
 
-import { environments } from './environments';
+import { rules } from './common/rules';
+import { findPort } from './common/find-ports';
 import { getVariables } from './runtime';
-import { app, rules } from './settings';
+import { app } from './settings';
+import { environments } from './environments';
 import { baseConfig } from './base.config';
-import { findPort } from './find-ports';
 
 const { style, compileExclusions } = app;
 const environmentVariables = getVariables(environments.local);
