+++ config/settings/index.js
@@ -1,7 +1,4 @@
 export * from './app';
 export * from './favicon';
-export * from './manifest';
 export * from './page';
 export * from './paths';
-export * from './project';
-export * from './rules';
